+++ app/wppwebjs/session/health.ts
@@ -2,7 +2,7 @@ import type { WAState } from "whatsapp-web.js";
 import { HEALTH_CHECK_INTERVAL_MS } from "./config";
 import type { HealthResult } from "../interfaces";
 import { safeDestroyClient, scheduleReconnect } from "./handlers";
-import { sessions } from "./sessions";
+import { sessions } from ".";
 import { raceWithTimeout } from "../../utils";
 
 export async function verifyClientHealth(companySlug: string): Promise<HealthResult> {
