--- app/wppwebjs/session/sessions.ts
@@ -1,3 +1,0 @@
-import type { SessionState } from "../interfaces";
-
-export const sessions: Record<string, SessionState> = {};
