--- app/wppwebjs/session/utils.ts
@@ -1,7 +1,0 @@
-import { sessions } from "./sessions";
-
-export function markSessionAsReady(companySlug: string) {
-    sessions[companySlug].ready = true;
-    sessions[companySlug].connecting = false;
-    sessions[companySlug].qrCode = null;
-}
